# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tecnofit API",
        "description": "API para gerenciamento de contas, saques e autenticação de usuários",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9501",
            "description": "Servidor local de desenvolvimento"
        }
    ],
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Contas"
                ],
                "summary": "Listar todas as contas",
                "description": "Retorna uma lista de todas as contas cadastradas",
                "operationId": "36472b568084733748dd5af6019d7752",
                "responses": {
                    "200": {
                        "description": "Lista de contas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "balance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 100
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Criar nova conta",
                "description": "Cria uma nova conta com saldo inicial",
                "operationId": "5bd68ba1ece64f7e05c637dfb97a573d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "balance"
                                ],
                                "properties": {
                                    "balance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "balance": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/accounts/{accountId}/balance/withdraw": {
            "post": {
                "tags": [
                    "Saques"
                ],
                "summary": "Realizar saque",
                "description": "Realiza um saque de uma conta. Pode ser saque imediato ou agendado.",
                "operationId": "6b854839a0250902bda7bac8612cb175",
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "description": "ID da conta",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "method",
                                    "pix"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Valor do saque",
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "method": {
                                        "description": "Método de saque",
                                        "type": "string",
                                        "enum": [
                                            "pix"
                                        ],
                                        "example": "pix"
                                    },
                                    "pix": {
                                        "properties": {
                                            "type": {
                                                "description": "Tipo de chave PIX",
                                                "type": "string",
                                                "enum": [
                                                    "cpf",
                                                    "cnpj",
                                                    "email",
                                                    "phone",
                                                    "random"
                                                ],
                                                "example": "cpf"
                                            },
                                            "key": {
                                                "description": "Chave PIX",
                                                "type": "string",
                                                "example": "12345678900"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "schedule": {
                                        "description": "Data e hora do agendamento (opcional, formato: Y-m-d H:i)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-31 14:30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Saque realizado ou agendado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Saque realizado com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "withdraw_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "scheduled": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "scheduled_for": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                },
                                                "pix_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "new_balance": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou saldo insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Erro ao processar requisição"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "Mensagem de erro"
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "O campo e-mail é obrigatório."
                            ],
                            "password": [
                                "O campo senha deve ter no mínimo 6 caracteres."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}